requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.7.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==41.0.7
dbus-python==1.3.2
discord.py==2.5.2
distro==1.9.0
frozenlist==1.5.0
httplib2==0.20.4
idna==3.10
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lxml==5.3.2
multidict==6.3.2
oauthlib==3.2.2
propcache==0.3.1
pycparser==2.22
PyGObject==3.48.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==68.1.2
six==1.16.0
soupsieve==2.6
ssh-import-id==5.11
typing_extensions==4.13.1
urllib3==2.3.0
wadllib==1.3.6
wheel==0.42.0
yarl==1.19.0